@@ -1,9 +1,8 @@
-// +build integration
+//go:build integration
 
 package test
 
 import (
-	"io/ioutil"
 	"os"
 	"testing"
 
@@ -22,14 +21,22 @@ func TestMultiClusterCase(t *testing.T) {
 		t.Error(err)
 		return
 	}
-	defer testenv.Environment.Stop()
+	t.Cleanup(func() {
+		if err := testenv.Environment.Stop(); err != nil {
+			t.Error(err)
+		}
+	})
 
 	testenv2, err := testutils.StartTestEnvironment(testutils.APIServerDefaultArgs, false)
 	if err != nil {
 		t.Error(err)
 		return
 	}
-	defer testenv2.Environment.Stop()
+	t.Cleanup(func() {
+		if err := testenv2.Environment.Stop(); err != nil {
+			t.Error(err)
+		}
+	})
 
 	podSpec := map[string]interface{}{
 		"restartPolicy": "Never",
@@ -41,12 +48,14 @@ func TestMultiClusterCase(t *testing.T) {
 		},
 	}
 
-	tmpfile, err := ioutil.TempFile("", "kubeconfig")
+	tmpfile, err := os.CreateTemp("", "kubeconfig")
 	if err != nil {
 		t.Error(err)
 		return
 	}
-	defer os.Remove(tmpfile.Name())
+	t.Cleanup(func() {
+		os.Remove(tmpfile.Name())
+	})
 	if err := testutils.Kubeconfig(testenv2.Config, tmpfile); err != nil {
 		t.Error(err)
 		return
