@@ -1,4 +1,4 @@
-// +build integration
+//go:build integration
 
 package test
 
@@ -9,6 +9,7 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	harness "github.com/kudobuilder/kuttl/pkg/apis/testharness/v1beta1"
+	testutils "github.com/kudobuilder/kuttl/pkg/test/utils"
 )
 
 func TestHarnessRunIntegration(t *testing.T) {
@@ -25,6 +26,30 @@ func TestHarnessRunIntegration(t *testing.T) {
 	harness.Run()
 }
 
+func TestHarnessRunIntegrationWithConfig(t *testing.T) {
+	testenv, err := testutils.StartTestEnvironment(nil, false)
+	if err != nil {
+		t.Fatalf("fatal error starting environment: %s", err)
+	}
+	config := &harness.RestConfig{RC: testenv.Config}
+	harness := Harness{
+		TestSuite: harness.TestSuite{
+			TestDirs: []string{
+				"./test_data/",
+			},
+			// set as true to skip service account check
+			StartControlPlane: true,
+			Config:            config,
+			CRDDir:            "./test_crds/",
+		},
+		T: t,
+	}
+	harness.Run()
+	if err := testenv.Environment.Stop(); err != nil {
+		t.Log("error tearing down mock control plane", err)
+	}
+}
+
 // This test requires external KinD support to run thus is an integration test
 func TestRunBackgroundCommands(t *testing.T) {
 	h := Harness{
@@ -38,7 +63,7 @@ func TestRunBackgroundCommands(t *testing.T) {
 	h.TestSuite.Commands = commands
 
 	h.Setup()
-	defer h.Stop()
+	t.Cleanup(h.Stop)
 
 	// setup creates bg processes
 	assert.Equal(t, 1, len(h.bgProcesses))
