@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"context"
 	"errors"
-	"io/ioutil"
 	"os"
 	"testing"
 	"time"
@@ -81,7 +80,7 @@ func TestGETAPIResource(t *testing.T) {
 	assert.Nil(t, err)
 	assert.Equal(t, apiResource.Kind, "Pod")
 
-	apiResource, err = GetAPIResource(fake, schema.GroupVersionKind{
+	_, err = GetAPIResource(fake, schema.GroupVersionKind{
 		Kind:    "NonExistentResourceType",
 		Version: "v1",
 	})
@@ -121,7 +120,6 @@ func TestRetryWithUnexpectedError(t *testing.T) {
 }
 
 func TestKubeconfigPath(t *testing.T) {
-
 	tests := []struct {
 		name     string
 		path     string
@@ -170,11 +168,11 @@ func TestRetryWithTimeout(t *testing.T) {
 }
 
 func TestLoadYAML(t *testing.T) {
-	tmpfile, err := ioutil.TempFile("", "test.yaml")
+	tmpfile, err := os.CreateTemp("", "test.yaml")
 	assert.Nil(t, err)
 	defer tmpfile.Close()
 
-	err = ioutil.WriteFile(tmpfile.Name(), []byte(`
+	err = os.WriteFile(tmpfile.Name(), []byte(`
 apiVersion: v1
 kind: Pod
 metadata:
@@ -246,11 +244,11 @@ spec:
 }
 
 func TestMatchesKind(t *testing.T) {
-	tmpfile, err := ioutil.TempFile("", "test.yaml")
+	tmpfile, err := os.CreateTemp("", "test.yaml")
 	assert.Nil(t, err)
 	defer tmpfile.Close()
 
-	err = ioutil.WriteFile(tmpfile.Name(), []byte(`
+	err = os.WriteFile(tmpfile.Name(), []byte(`
 apiVersion: v1
 kind: Pod
 metadata:
@@ -430,7 +428,6 @@ func TestGetKubectlArgs(t *testing.T) {
 		test := test
 
 		t.Run(test.testName, func(t *testing.T) {
-
 			if test.env != nil || len(test.env) > 0 {
 				for key, value := range test.env {
 					os.Setenv(key, value)
@@ -491,7 +488,6 @@ func TestRunScript(t *testing.T) {
 	}
 
 	for _, tt := range tests {
-
 		tt := tt
 
 		t.Run(tt.name, func(t *testing.T) {
