@@ -7,7 +7,6 @@ import (
 )
 
 func TestFromPath(t *testing.T) {
-
 	tests := []struct {
 		name     string
 		path     string
@@ -40,8 +39,8 @@ func TestFromPath(t *testing.T) {
 
 	for _, tt := range tests {
 		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
 
+		t.Run(tt.name, func(t *testing.T) {
 			paths, err := FromPath(tt.path, tt.pattern)
 			assert.Equal(t, tt.wantErr, err != nil, "expected error %v, but got %v", tt.wantErr, err)
 			assert.ElementsMatch(t, paths, tt.expected)
@@ -75,8 +74,8 @@ func TestTrimExt(t *testing.T) {
 
 	for _, tt := range tests {
 		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
 
+		t.Run(tt.name, func(t *testing.T) {
 			path := TrimExt(tt.path)
 			assert.Equal(t, path, tt.expected)
 		})
